app/requirements.txt
Flask
google-cloud-datastore
google-auth
requests
flask-login
google-api-python-client
